migração definitiva para json
