--- extrato.txt
@@ -1,55 +1,0 @@
-[2025-06-18 20:42:54] TRANSFERÊNCIA: R$56.00
-[2025-06-18 20:44:28] DEPÉSITO: R$56.00
-[2025-06-18 20:44:58] DEPÉSITO: R$567.00
-[2025-06-18 20:48:12] DEPÉSITO: R$56.00
-[2025-06-18 20:48:54] DEPÉSITO: R$567.00
-[2025-06-18 20:49:04] DEPÉSITO: R$6789.00
-[2025-06-18 20:51:16] DEPÉSITO: R$3.00
-[2025-06-18 20:52:05] DEPÉSITO: R$456.00
-[2025-06-18 20:52:44] DEPÉSITO: R$456.00
-[2025-06-18 20:54:13] DEPÉSITO: R$4667.00
-[2025-06-18 20:55:49] DEPÉSITO: R$56.00
-[2025-06-18 20:56:26] DEPÉSITO: R$56.00
-[2025-06-18 20:56:36] DEPÉSITO: R$35.00
-[2025-06-18 20:56:52] DEPÉSITO: R$67.00
-[2025-06-18 20:57:01] DEPÉSITO: R$34.00
-[2025-06-18 20:57:43] DEPÉSITO: R$56.00
-[2025-06-18 20:58:21] DEPÓSITO: R$34.00
-[2025-06-18 20:58:49] DEPÓSITO: R$56.00
-[2025-06-18 20:59:29] DEPÓSITO: R$76.00
-[2025-06-18 20:59:48] DEPÓSITO: R$345.00
-[2025-06-18 21:01:11] TRANSFERÊNCIA: R$567.00
-[2025-06-18 21:01:19] TRANSFERÊNCIA: R$678.00
-[2025-06-18 21:04:25] DEPÓSITO: R$6740.00
-[2025-06-18 21:04:33] TRANSFERÊNCIA: R$5679.00
-[2025-06-18 21:16:02] DEPÓSITO: R$45.00
-[2025-06-18 21:16:20] TRANSFERÊNCIA: R$1232.00
-[2025-06-19 17:53:38] DEPÓSITO: R$ 567.00
-[2025-06-19 17:54:09] DEPÓSITO: R$ 456.00
-[2025-06-19 17:54:33] DEPÓSITO: R$ 1.00
-[2025-06-19 17:55:33] DEPÓSITO: R$ 5.00
-[2025-06-19 17:59:18] DEPÓSITO: R$ 233.00
-[2025-06-19 18:02:55] DEPÓSITO: R$ 345.00
-[2025-06-19 18:04:22] DEPÓSITO: R$ 345.00
-[2025-06-19 18:05:33] DEPÓSITO: R$ 3454.00
-[2025-06-19 18:06:26] DEPÓSITO: R$ 67.00
-[2025-06-19 18:08:50] DEPÓSITO: R$ 3.00
-[2025-06-19 18:21:23] DEPÓSITO: R$ 567.00
-[2025-06-19 18:21:33] TRANSFERÊNCIA: R$ 345.00
-[2025-06-19 19:16:18] DEPÓSITO: R$ 345.00
-[2025-06-19 19:16:50] TRANSFERÊNCIA: R$ 56.00
-[2025-06-19 19:34:21] DEPÓSITO: R$ 45.00
-[2025-06-19 20:00:58] PAGAMENTOS: R$ 456.00
-[2025-06-19 20:09:22] DEPÓSITO: R$ 5678.00
-[2025-06-19 20:09:53] TRANSFERÊNCIA: R$ 455.00
-[2025-06-19 20:13:30] PAGAMENTOS: R$ 5.00
-[2025-06-19 20:45:36] DEPÓSITO: R$ 3456.00
-[2025-06-19 20:49:39] TRANSFERÊNCIA: R$ 7900.00
-[2025-06-19 20:52:11] DEPÓSITO: R$ 577.00
-[2025-06-19 20:53:55] TRANSFERÊNCIA: R$ 638.00
-[2025-06-19 21:18:43] TRANSFERÊNCIA: R$ 577.00
-[2025-06-19 21:22:04] DEPÓSITO: R$ 44.00
-[2025-06-19 21:22:11] TRANSFERÊNCIA: R$ 2445.00
-[2025-06-19 21:28:29] DEPÓSITO: R$ 444.00
-[2025-06-19 22:13:23] PAGAMENTOS: R$ 3333.00
-[2025-06-19 22:33:35] DEPÓSITO: R$ 6899303837447485.00
