+++ main.py
@@ -1,4 +1,3 @@
-#from calculadora import escolher_transacao
 from calculadora import escolher_transacao
 
 if __name__ == "__main__":
